+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityChiseledBookshelf.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.blockentity.interfaces;
+
+import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.component.BlockEntityChiseledBookshelfContainerHolderComponent;
+
+/**
+ * @author daoge_cmd
+ */
+public interface BlockEntityChiseledBookshelf extends BlockEntity, BlockEntityChiseledBookshelfContainerHolderComponent {
+}